requirements.txt
gradio
numpy
torch
torchvision
matplotlib
pillow
opencv-python